Indicate the record of the Ten Tribes
Create a node showing when the Ten Tribes were carried away into the
land of Assyria.  We won't be able to follow that edge very far since
we don't know what happened to them after that.  Nevertheless, when
the record of the Ten Tribes is discovered, we'll be able to extend
the graph and use this node to verify the confirm the authenticity of
the record.