Document the first quote from the Brass Plates
This new Turtle RDF file presents a first quote from the Brass Plates
in the Book of Mormon.  It also presents a sample RDF schema for
describing quotes such as this.

Eventually, the Book of Mormon transmission graph should be encoded
using RDF as well.  That will let us specify where on the graph each
of these quotes occurs.  We can then triangulate in which section the
original source for each quote lies.  We can also generate the Book of
Mormon transmission chart from that data.