Encode events in Mercury
This is a first experiment with encoding scriptural events and
associated metadata in the Mercury programming language.  Compared to
experiments with RDF, Perl and Prolog, this is the most successful.
The graph-oriented data model as powerful as RDF and Prolog, but
having algebraic data types helps me keep all the identifiers
straight.  This data is going to get large, very quickly.  The Perl
experiment was only 172 lines and I already had trouble keeping things
straight.

This experiment also has the first representation of scripture
references that I'm happy with.  Again, algebraic data types to the
rescue.

Mercury's module system will let me keep the scripture reference
types, event types, people types, location types, etc. in separate
modules.  As the data grows larger and larger, this should be very
manageable.

To get a feel for querying the data, this program just displays all
events in which Lehi participated.  Mercury's mode system caught a few
errors in this simple problem.

The model should have enough detail to encode all the information
needed for tracing quotes back to the Brass Plates.  event/2 provides
nodes; and_then/3 provides edges.  Scripture passages can be located
along edges or nodes.  If I implement comparison operators for
scripture references, it should be easy enough to locate a given
scripture on an edge or node and then traverse and_then-parentage to
find the most recent, common ancestor.