Label edges in the graph image
Label edges with a succinct list of people who traveled that edge.
The little extra information makes it much easier to understand who
participated in each leg of the journey.

This code somehow suggested to me that events may not need to record
who/1 predicates.  We might be able to infer that information based
only on lists of people participating in the edges.  Although, there
may be cases I haven't encountered where that's not true.  I should
think about it some more.  If it works, it could be a useful
simplification.